Resolve BASIS-175045

Closes BASIS-175045

See merge request basis/abaco_codigo_fonte!112

+++ frontend/src/app/sistema/sistema-list/sistema-list.component.ts
@@ -1,11 +1,12 @@
-import {Component, ViewChild} from '@angular/core';
-import {Router} from '@angular/router';
-import {ConfirmationService} from 'primeng';
+import { Component, ViewChild } from '@angular/core';
+import { Router } from '@angular/router';
+import { ConfirmationService } from 'primeng';
 import { DatatableComponent, PageNotificationService, DatatableClickEvent } from '@nuvem/primeng-components';
 import { Sistema } from '../sistema.model';
 import { ElasticQuery } from 'src/app/shared/elastic-query';
 import { Organizacao, OrganizacaoService } from 'src/app/organizacao';
 import { SistemaService } from '../sistema.service';
+import { AuthService } from 'src/app/util/auth.service';
 
 @Component({
     selector: 'app-sistema',
@@ -18,7 +19,7 @@ export class SistemaListComponent {
     searchUrl: string = this.sistemaService.searchUrl;
     sistemaSelecionado: Sistema;
     rowsPerPageOptions: number[] = [5, 10, 20];
-    paginationParams = {contentIndex: null};
+    paginationParams = { contentIndex: null };
     elasticQuery: ElasticQuery = new ElasticQuery();
     organizations: Organizacao[] = [];
     customOptions: Object = {};
@@ -38,13 +39,14 @@ export class SistemaListComponent {
         private confirmationService: ConfirmationService,
         private organizacaoService: OrganizacaoService,
         private pageNotificationService: PageNotificationService,
+        private authService: AuthService
     ) {
         const emptyOrganization = new Organizacao();
         this.organizacaoService.dropDown().subscribe(response => {
 
             this.customOptions['organizacao'] = response.map((item) => {
-                return {label: item.nome, value: item.id};
-              });
+                return { label: item.nome, value: item.id };
+            });
             this.organizations = response;
             this.organizations.push(emptyOrganization);
         });
@@ -72,12 +74,18 @@ export class SistemaListComponent {
         }
         switch (event.button) {
             case 'edit':
+                if (this.authService.possuiRole(AuthService.PREFIX_ROLE + "SISTEMA_EDITAR") == false) {
+                    break;
+                }
                 this.router.navigate(['/sistema', event.selection.id, 'edit']);
                 break;
             case 'delete':
                 this.confirmDelete(event.selection.id);
                 break;
             case 'view':
+                if (this.authService.possuiRole(AuthService.PREFIX_ROLE + "SISTEMA_CONSULTAR") == false) {
+                    break;
+                }
                 this.router.navigate(['/sistema', event.selection.id]);
                 break;
         }
@@ -93,18 +101,24 @@ export class SistemaListComponent {
     }
 
     abrirEditar() {
+        if (this.authService.possuiRole(AuthService.PREFIX_ROLE + "SISTEMA_EDITAR") == false) {
+            return false;
+        }
         this.router.navigate(['/sistema', this.sistemaSelecionado.id, 'edit']);
     }
 
     public confirmDelete(id: any) {
+        if (this.authService.possuiRole(AuthService.PREFIX_ROLE + "SISTEMA_EXCLUIR") == false) {
+            return false;
+        }
         this.confirmationService.confirm({
             message: this.getLabel('Tem certeza que deseja excluir o registro?'),
             accept: () => {
                 this.sistemaService.delete(id).subscribe(() => {
-                        this.limparPesquisa();
-                        this.pageNotificationService.addDeleteMsg();
-                    }, (error) => {
-                    }
+                    this.limparPesquisa();
+                    this.pageNotificationService.addDeleteMsg();
+                }, (error) => {
+                }
                 );
             }
         });
@@ -137,20 +151,20 @@ export class SistemaListComponent {
         this.searchParams.nomeSistema = this.formatFieldForSearch(this.searchParams.nomeSistema);
         if (this.searchParams.sigla !== undefined && this.searchParams.sigla !== '') {
             if (this.searchParams.sigla.includes(' ')) {
-                stringParamsArray.length > 0 ? stringParamsArray.push(' AND sigla:\"' + this.searchParams.sigla + '\"') : stringParamsArray.push('sigla:\"' + this.searchParams.sigla + '\"'); 
+                stringParamsArray.length > 0 ? stringParamsArray.push(' AND sigla:\"' + this.searchParams.sigla + '\"') : stringParamsArray.push('sigla:\"' + this.searchParams.sigla + '\"');
             } else {
-                stringParamsArray.length > 0 ? stringParamsArray.push(' AND sigla:*' + this.searchParams.sigla + '*') :  stringParamsArray.push('sigla:*' + this.searchParams.sigla + '*'); 
+                stringParamsArray.length > 0 ? stringParamsArray.push(' AND sigla:*' + this.searchParams.sigla + '*') : stringParamsArray.push('sigla:*' + this.searchParams.sigla + '*');
             }
         }
         if (this.searchParams.nomeSistema !== undefined && this.searchParams.nomeSistema !== '') {
             if (this.searchParams.nomeSistema.includes(' ')) {
-                stringParamsArray.length > 0 ? stringParamsArray.push(' AND nome:\"' + this.searchParams.nomeSistema + '\"') : stringParamsArray.push('nome:\"' + this.searchParams.nomeSistema + '\"'); 
+                stringParamsArray.length > 0 ? stringParamsArray.push(' AND nome:\"' + this.searchParams.nomeSistema + '\"') : stringParamsArray.push('nome:\"' + this.searchParams.nomeSistema + '\"');
             } else {
-                stringParamsArray.length > 0 ? stringParamsArray.push(' AND nome:*' + this.searchParams.nomeSistema + '*') :  stringParamsArray.push(this.searchParams.nomeSistema); 
+                stringParamsArray.length > 0 ? stringParamsArray.push(' AND nome:*' + this.searchParams.nomeSistema + '*') : stringParamsArray.push(this.searchParams.nomeSistema);
             }
         }
         if (this.searchParams.organizacao.id !== undefined && this.searchParams.organizacao.id !== '') {
-                stringParamsArray.length > 0 ? stringParamsArray.push(' AND organizacao.id: '+ this.searchParams.organizacao.id ) : stringParamsArray.push(' organizacao.id:' + this.searchParams.organizacao.id); 
+            stringParamsArray.length > 0 ? stringParamsArray.push(' AND organizacao.id: ' + this.searchParams.organizacao.id) : stringParamsArray.push(' organizacao.id:' + this.searchParams.organizacao.id);
         }
         return stringParamsArray;
     }
@@ -165,39 +179,50 @@ export class SistemaListComponent {
     }
 
     public recarregarDataTable() {
+        if (this.authService.possuiRole(AuthService.PREFIX_ROLE + "SISTEMA_PESQUISAR") == false) {
+            return false;
+        }
         this.datatable.url = this.searchUrl;
         this.datatable.refresh(this.elasticQuery.value ? this.elasticQuery.value : this.elasticQuery.query);
     }
 
 
-  concatResults(paramsArray: Array<string>): string {
-    let paramsQueue: Array<string> = [];
+    concatResults(paramsArray: Array<string>): string {
+        let paramsQueue: Array<string> = [];
 
-    if (paramsArray) {
-      paramsArray.forEach(each => {
-        (each !== undefined) ? (paramsQueue.push(each)) : (each);
-      });
+        if (paramsArray) {
+            paramsArray.forEach(each => {
+                (each !== undefined) ? (paramsQueue.push(each)) : (each);
+            });
+        }
+
+        const concatResultString = this.createString(paramsQueue);
+
+        return concatResultString;
     }
 
-    const concatResultString = this.createString(paramsQueue);
+    private createString(paramsQueue: Array<String>): string {
+        let concatedString = '';
 
-    return concatResultString;
-  }
-
-  private createString(paramsQueue: Array<String>): string {
-    let concatedString = '';
-
-    for (let i = 0 ; i < paramsQueue.length ; i ++) {
-      (i !== 0) ? (concatedString = concatedString + ' + ' + paramsQueue[i]) : (concatedString = concatedString + paramsQueue[0]);
+        for (let i = 0; i < paramsQueue.length; i++) {
+            (i !== 0) ? (concatedString = concatedString + ' + ' + paramsQueue[i]) : (concatedString = concatedString + paramsQueue[0]);
+        }
+        return concatedString;
     }
-    return concatedString;
-  }
 
-  public selectTipoEquipe() {
-    if (this.datatable && this.datatable.selectedRow) {
-        if (this.datatable.selectedRow && this.datatable.selectedRow) {
-            this.sistemaSelecionado = this.datatable.selectedRow;
-          }
-      }
-  }
+    public selectTipoEquipe() {
+        if (this.datatable && this.datatable.selectedRow) {
+            if (this.datatable.selectedRow && this.datatable.selectedRow) {
+                this.sistemaSelecionado = this.datatable.selectedRow;
+            }
+        }
+    }
+
+    criarSistema(){
+        if (this.authService.possuiRole(AuthService.PREFIX_ROLE + "SISTEMA_CADASTRAR") == false) {
+            return false;
+        }
+
+        this.router.navigate(["/sistema/new"]);
+    }
 }
